moving from deprecated v13.app.FragmentStatePagerAdapter to v4.app.FragmentStatePagerAdapter as recommended
