+++ WordPress/src/main/java/org/wordpress/android/ui/comments/CommentDetailFragment.java
@@ -2,14 +2,14 @@ package org.wordpress.android.ui.comments;
 
 import android.app.Activity;
 import android.app.AlertDialog;
-import android.app.Fragment;
-import android.app.FragmentManager;
-import android.app.FragmentTransaction;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentTransaction;
 import android.support.v4.content.ContextCompat;
 import android.text.Html;
 import android.text.TextUtils;
@@ -691,7 +691,7 @@ public class CommentDetailFragment extends Fragment implements NotificationFragm
             }
         }
 
-        getFragmentManager().invalidateOptionsMenu();
+        getActivity().invalidateOptionsMenu();
     }
 
     /*
@@ -1103,7 +1103,7 @@ public class CommentDetailFragment extends Fragment implements NotificationFragm
 
         addDetailFragment(note.getId());
 
-        getFragmentManager().invalidateOptionsMenu();
+        getActivity().invalidateOptionsMenu();
     }
 
     private void addDetailFragment(String noteId) {
