+++ WordPress/src/main/java/org/wordpress/android/ui/comments/CommentDetailFragmentAdapter.java
@@ -1,9 +1,9 @@
 package org.wordpress.android.ui.comments;
 
-import android.app.Fragment;
-import android.app.FragmentManager;
 import android.os.Parcelable;
-import android.support.v13.app.FragmentStatePagerAdapter;
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentStatePagerAdapter;
 
 import org.wordpress.android.fluxc.model.CommentModel;
 import org.wordpress.android.fluxc.model.SiteModel;
